De nuevo

+++ visita-facil-api/src/main/java/com/construsoft/visita_facil_api/domain/SolicitudVisitaDTO.java
@@ -14,7 +14,7 @@ public class SolicitudVisitaDTO {
     private String correoCliente;
     private String telefonoCliente;
 
-    public Integer getPropiedadId() {
+    public Long getPropiedadId() {
         return propiedadId;
     }
